Add support to the subunit workers to reuse zuul uuids
The latest subunit2sql release, 0.2.1, added support for setting the
id for a run. This commit adds support to the gearman workers for
doing this using the zuul build_uuids. This way if an event has a
build_uuid set in the metadata the id column for the row will match
zuul's. This will be useful for when eventually want to correlate
data between zuul and subunit2sql.

Change-Id: Ic4aa13a6c92c3bc759e080fdc4c2e07ff8b881bd