Disable cgit's dumb HTTP protocol endpoint support
* modules/openstack_project/files/git/cgitrc: The dumb HTTP protocol
support in cgit is an attractive nuisance and source of surprising
random errors for users who misconfigure their repositories to use
it as a remote. We provide a separate smart HTTP backend and
configure cgit to list that as the clone URL anyway, so the default
enable-http-clone is disabled to avoid further accidental use.

Change-Id: Iacc34ba51cd835f7e999b753e06ef6eea41630df